requirements.txt
uvicorn==0.30.6
fastapi==0.114.2
openai==1.45.0
langchain
langchain-community==0.3.0
chromadb
llama-index-core
llama-parse
llama-index-readers-file
python-dotenv
llama-index-readers-google==0.4.0
langchain
langchain-community==0.3.0
chromadb
llama-index-core
llama-parse
llama-index-readers-file
python-dotenv
llama-index-readers-google==0.4.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain-google-community[drive]
google-api-python-client-stubs